pyproject.toml
[project]
name = "llmops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "langchain-openai>=0.3.31",
    "langsmith>=0.4.16",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
